Fix build can not able to be bootstrapped
If you try to build topoprocessing with a fresh Maven local repo the
build will fail because it cannot retrieve the parent pom.xml file. This
is because we need to define a URL to retrieve it. The current method of
retrieving the URL via <nexusproxy> variable does not work until the
parent pom.xml (odlparent) can be downloaded first.

This patch removes the unnecessary declaration of <pluginRepositories>
and <distributionManagement> sections since they are inherited from the
parent pom.xml. The patch also fixes the <repositories> definition so
that the repo build can be bootstrapped with a fresh Maven local repo.

Change-Id: If44974b8bbefaf79beaebe9d3f79c72ff7b93a5d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>