Update patch set 1
Patch Set 1:

(3 comments)

The main hurdle seem to be that the state with no work pending has two different symptoms: Empty queue and writeScheduled zero; and it is not easy to block for both of them at once.

Currently I think that instead of preparedOperations, the "blockably shutdownable" object should by made of pool, writeScheduled and shutdown flag. And the final deleteOperation should be directly submitted to transactionChain after the pool is confirmed to be shut down (instead of using queue+pool mechanism for submitting the operation).

Patch-set: 1
Label: Verified=0