Update patch set 2
Patch Set 2:

Hi Icaro, we planed to remove topoprocessing-common module, because we are not using it. Could you please provide us some information what purpose your change has? We can then decide together, if we can keep it in topoprocessing-common module, add it to some another module, or if it will be better to create new module for your functionality. Thanks, Andrej...

Patch-set: 2