Update patch set 5
Patch Set 5:

Sure, we can wait. He warned me he'll be back early December, so it can wait. By then, I might rework the patch to avoid having to save all the reg in the RpcServices as i've done, and have them held and closed in the class registering the services. TBH I'm not keen of what I've done in RpcServices (it's more a quick hack than a nice solution)

Patch-set: 5
Label: Verified=0