Update patch set 5
Patch Set 5:

I looked to your request and I want to point out one thing about used script.
```
<aggregation>
    <aggregation-type>equality</aggregation-type>
    <scripting>
        <language>javascript</language>
        <script> if ( newItem.getItemId().replace(/domain=(.*&amp;)/,"")==originalItem.getItemId().replace(/domain=(.*&amp;)/,"")) {aggregable.setResult(true);} else {aggregable.setResult(false);} </script>
    </scripting>
    <mapping>
        <input-model>network-topology-model</input-model>
        <underlay-topology>bgp-example1</underlay-topology>
        <target-field>
            <target-field-path>network-topology:link-id</target-field-path>
            <matching-key>0</matching-key>
        </target-field>
        <aggregate-inside>false</aggregate-inside>
    </mapping>
    <mapping>
        <input-model>network-topology-model</input-model>
        <underlay-topology>bgp-example2</underlay-topology>
        <target-field>
            <target-field-path>network-topology:link-id</target-field-path>
            <matching-key>0</matching-key>
        </target-field>
        <aggregate-inside>false</aggregate-inside>
    </mapping>
</aggregation>
```

in script you have `newItem.getItemId()` I don't know if it was intended, but I have to point out, that this itemId may not correspond with link-id which you have in mapping. If you want to work with link id which you have in mapping, than in script you should use this `newItem.getLeafNodes().get(<matching_key>).getValue()`. The same for `originalItem`.
So if your script works for you, I assume it is only by coincidence that during processing of link we use link-id as itemId for UnderlayItem. In other words, I think it will not work for target fields different as link-id.

This makes me believe, that we should change scripting in a way, that only target fields will be accesible, not whole UnderlayItem. But it is for another change/bug/feature.

Patch-set: 5
Reviewer: Gerrit User 4630 <4630@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0